Struct cudnn::utils::ConvolutionConfig

pub struct ConvolutionConfig {
    // some fields omitted
}

Provides a convenient interface to access cuDNN's convolution parameters, algo and workspace and workspace_size_in_bytes.

You woudn't use this struct yourself, but rather obtain it through Cudnn.init_convolution().

Methods

impl ConvolutionConfig

fn new(algo_fwd: cudnnConvolutionFwdAlgo_t, workspace_size_fwd: usize, algo_filter_bwd: cudnnConvolutionBwdFilterAlgo_t, workspace_filter_size_bwd: usize, algo_data_bwd: cudnnConvolutionBwdDataAlgo_t, workspace_data_size_bwd: usize, conv_desc: ConvolutionDescriptor, filter_desc: FilterDescriptor) -> ConvolutionConfig

Returns a new ConvolutionConfig

fn largest_workspace_size(&self) -> &usize

Returns the largest workspace size out of the three.

Useful for creating a shared workspace.

fn forward_algo(&self) -> &cudnnConvolutionFwdAlgo_t

Returns forward_algo.

fn forward_workspace_size(&self) -> &usize

Returns forward_workspace_size.

fn backward_filter_algo(&self) -> &cudnnConvolutionBwdFilterAlgo_t

Returns backward_filter_algo.

fn backward_filter_workspace_size(&self) -> &usize

Returns backward_filter_workspace_size.

fn backward_data_algo(&self) -> &cudnnConvolutionBwdDataAlgo_t

Returns backward_data_algo.

fn backward_data_workspace_size(&self) -> &usize

Returns backward_data_workspace_size.

fn conv_desc(&self) -> &ConvolutionDescriptor

Returns conv_desc.

fn filter_desc(&self) -> &FilterDescriptor

Returns filter_desc.